Fixed a type conversion warning

+++ vertexarray.cpp
@@ -37,13 +37,13 @@ void VertexArray::AddBuffer(const VertexBuffer& vb, const VertexBufferLayout &la
     Bind();
     vb.Bind();
     const auto& elements = layout.GetElements();
-    unsigned long offset = 0;
+    unsigned int offset = 0;
     for ( unsigned int ii = 0; ii < elements.size(); ++ii )
     {
         const auto& element = elements[ii];
         glEnableVertexAttribArray(ii);
         glVertexAttribPointer(ii, element.count, element.type, element.normalized,
-                layout.GetStride(), (const void*)offset);
+                layout.GetStride(), (const void*)(uint64_t)offset);
         offset += element.count * VertexBufferElement::GetSizeOfType(element.type);
     }
 }
